backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astroid==2.15.8
autopep8==2.0.4
cachetools==5.3.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.8
distro==1.9.0
Django==4.2.10
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.7.0
django-filter==23.2
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
filelock==3.13.1
fsspec==2024.2.0
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-speech==2.24.1
google-cloud-texttospeech==2.16.1
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
grpcio==1.60.1
grpcio-status==1.60.1
gTTS==2.5.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==1.26.4
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pylint==2.17.4
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.27.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
sniffio==1.3.0
sqlparse==0.4.4
tokenizers==0.15.2
tomlkit==0.12.3
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
wrapt==1.16.0
scikit-learn==1.2.2
